Update user's UID/GID on container startup
The handoff manager in the Docker container must be run by a user having
read/write access to the buffer and the storage area.  In previous
approach, UID/GID of that user was being persisted in the image itself.
As a result Docker image with the manager built on one machine could not
be used on other.  To address this limitation I rewrote Docker
configuration files to allow for dynamic updating of user's UID/GID
during container startup.